sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moltbridge"
version = "0.1.0"
description = "Python SDK for MoltBridge — professional network intelligence for AI agents"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "SageMind AI", email = "support@moltbridge.com" },
]
keywords = ["ai", "agents", "networking", "graph", "moltbridge"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "httpx>=0.27.0",
  "PyNaCl>=1.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-httpx>=0.30",
  "respx>=0.21",
]

[project.urls]
Homepage = "https://moltbridge.com"
Documentation = "https://docs.moltbridge.com"
Repository = "https://github.com/SageMindAI/moltbridge-python"
